fix: 修复编译错误
- 修复 CameraMapping.kt 的 pow 函数调用参数类型
- 修复 YanbaoMemoryManager.kt 的 YanbaoMemory 构造函数调用
- 添加 Flow 导入
- 修复 ControlPanel.kt 的 clickable 使用错误和缩进问题

+++ app/src/main/java/com/yanbao/camera/core/util/CameraMapping.kt
@@ -50,7 +50,7 @@ object CameraMapping {
     fun mapAngleToISO(angle: Float): Int {
         val progress = angle / 360f
         // 对数映射：100, 200, 400, 800, 1600, 3200, 6400
-        val logValue = 100 * (2.0.pow(progress * 6))
+        val logValue = 100 * (2.0.pow((progress * 6).toDouble()))
         return logValue.toInt().coerceIn(100, 6400)
     }
